Now place PDFs into a subdirectory of the location
Previously, generated PDF files were placed into a special location
off of the root, aptly named 'pdf/'. This led to collisions when
deploying more than one publishing setup on the same server. Now,
generated PDFs are placed into a sublocation of the current location,
so only files in the sublocation owned by this project are
overwritten during a deployment operation.